Remove Error.logs
This is currently not used in the Java agent
Also, the fact that we are always instantiate Logs
leads to validation errors in the APM server,
as Logs.message is required.

The reason we always instantiate Logs is that
it is required for recycling the Error object.

A future solution would be to have a boolean flag on Logs
to indicate if it is actually populated.
If not, render as null in the JSON representation.

Signed-off-by: Felix Barnsteiner <felix.barnsteiner@elastic.co>