Deploy war file with copyFileToContainer
This works around a bug in docker-java which encodes the plus sign to
a space when binding a file to the container.

Also restructures the integration tests into a top-level
integration-tests module as it felt a bit awkward as a apm-agent-plugins
submodule. It's not a plugin and we want to avoid packaging the tests
applications like simple-webapp into the main distribution.

Signed-off-by: Felix Barnsteiner <felix.barnsteiner@elastic.co>