Correct handling of NOTICE and LICENSE file in distribution
The agent heavily relies on shading aka. relocating of dependencies
to avoid version conflicts. In legal terms that means we are bundling
3rd party software.

# LICENSE file

According to
http://www.apache.org/dev/licensing-howto.html#permissive-deps,
the LICENSE file has to contain pointers to non-Apache 2.0 licenses of
bundled dependencies and has to be included in binary releases.
The proper location for that is META-INF/LICENSE.

# NOTICE file

http://www.apache.org/dev/licensing-howto.html#mod-notice states:
> It is important to keep NOTICE as brief and simple as possible,
as each addition places a burden on downstream consumers.

> Do not add anything to NOTICE which is not legally required.

Just like the LICENSE file, the NOTICE should also be contained in the
binary release under META-INF/NOTICE.

In the Apache commons-lang library, for example, there is a notice about
a method copied from the Spring Framework:
https://github.com/apache/commons-lang/blob/master/NOTICE.txt#L7

Signed-off-by: Felix Barnsteiner <felix.barnsteiner@elastic.co>