Make sure the error context is always set
Previously, most of the transaction context was set after the execution of a request.
When capturing an exception, the transaction context is copied over at it's current state.
That means when the user captures an exception, the the transaction context is not yet
populated and therefore, the Error context is empty.

This changes the initialization of the transaction context so that it is set
before the request starts.

That way, we can be sure that the Error context is always set