Allow instrumentations to have multiple group names
This, for example, enables to disable only the async part of the
Servlet API to be disabled via `servlet-api-async`, or to disable all
the Servlet API instrumentation via `servlet-api`.

Disables `incubating` by default.
This allows to release incubating instrumentations without risk but
makes it very easy for early adopters to test an incubating
instrumentation.