Don't limit size of caches
The size can only be limited on a per-class loader basis.
There can be lots of class loaders, especially in OSGi environments
so limiting the cache size is not a sufficient guard against OOMEs.
Also, limiting based on cache size does not take into account wether or
not a TypeDescription is lazy (only holds the class name) or has already
been resolved, which makes a big difference in size.

Instead, use SoftReferences to guard against OOMEs.