Switch from Java 8 interfaces to custom interfaces
Those can also be implemented via a lambda expression but don't have a
Java 7/8 compatibility issue.

The interfaces are invoked via a MethodHandle which is constructed in
the API and read from method arguments in the instrumentation.

Therefore, the instrumentation does not have to dynamically create
MethodHandle or Method objects.