Update render config docs when testing
Tests will still fail on CI if the docs are not up-to-date.
But you can bring them up-to-date by executing the tests.
The advantage is that you don't have to execute a main method from within the IDE.
Executing the main method via maven is also not trivial for multi-module maven projects.