Fix performance regression in indy servlet plugin
Reduces allocations by 32b/op back to close to 0,
which is where it was before migrating the Servlet plugin to an indy
plugin.
The issue is that DetachedThreadLocal#put allocates a new key even if
the key is already present in the map.
Also, using a default value allocates an entry on lookup.