Replace System Rules with System Lambda
System Lambda is independent of the test framework while System Rules is
build for JUnit 4. We are using JUnit Jupiter and therefore System
Lambda is the better choice.

In addition System Lambda is more precise. It allows to set environment
variables for a single statement only. This already improves the
readability of the test because you can now immediately see which
environment variables are set when SystemInfo#findContainerDetails is
executed.