Reduce GC time overhead caused by WeakReferences
This is achieved by reducing the live set of WeakReferences in JDBC plugin.

After a Statement has been closed, it's safe to remove it from the Statement/String map as it can't be used again.
This reduces the number of weak references to Statements in case the app/framework properly closes statements.

This is significant because WeakReferences increase the duration of the `Notify Soft/WeakReferences` GC phase.